+++ src/main/java/com/scubakay/zombiescantgather/command/RootCommand.java
@@ -1,6 +1,7 @@
 package com.scubakay.zombiescantgather.command;
 
 import com.mojang.brigadier.CommandDispatcher;
+import com.mojang.brigadier.tree.CommandNode;
 import com.mojang.brigadier.tree.LiteralCommandNode;
 import net.minecraft.command.CommandRegistryAccess;
 import net.minecraft.server.command.CommandManager;
@@ -8,18 +9,37 @@ import net.minecraft.server.command.ServerCommandSource;
 
 public class RootCommand {
     public static String ROOT_COMMAND = "zombiescantgather";
-    private static LiteralCommandNode<ServerCommandSource> root;
+    public static String ALIAS_COMMAND = "zcg";
+    private final LiteralCommandNode<ServerCommandSource> root;
+    private final LiteralCommandNode<ServerCommandSource> alias;
+    private static RootCommand command;
 
-    public static void register(CommandDispatcher<ServerCommandSource> dispatcher, CommandRegistryAccess ignoredCommandRegistryAccess, CommandManager.RegistrationEnvironment ignoredRegistrationEnvironment) {
-        dispatcher.getRoot().addChild(getRoot());
+    public RootCommand(CommandDispatcher<ServerCommandSource> dispatcher) {
+        root = CommandManager
+                .literal(ROOT_COMMAND)
+                .build();
+        alias = CommandManager
+                .literal(ALIAS_COMMAND)
+                .build();
+        dispatcher.getRoot().addChild(root);
+        dispatcher.getRoot().addChild(alias);
     }
 
-    protected static LiteralCommandNode<ServerCommandSource> getRoot() {
-        if (root == null) {
-            root = CommandManager
-                    .literal(ROOT_COMMAND)
-                    .build();
+    public static RootCommand getRoot(CommandDispatcher<ServerCommandSource> dispatcher) {
+        if (command == null) {
+            command = new RootCommand(dispatcher);
         }
-        return root;
+        return command;
+    }
+
+    public void addChild(CommandNode<ServerCommandSource> node) {
+        this.root.addChild(node);
+        this.alias.addChild(node);
+    }
+
+    public static void register(CommandDispatcher<ServerCommandSource> dispatcher, CommandRegistryAccess ignoredCommandRegistryAccess, CommandManager.RegistrationEnvironment ignoredRegistrationEnvironment) {
+        if (command == null) {
+            command = new RootCommand(dispatcher);
+        }
     }
 }